Revert "limit number of answers a user can see (#2356)" (#2361)
This reverts commit ba1c29d17ccd6d632161554fb201f831c3572b11.
