+++ src/apps/homework/api/filtersets.py
@@ -1,6 +1,6 @@
 from django_filters import rest_framework as filters
 
-from apps.homework.models import Answer, AnswerCrossCheck
+from apps.homework.models import Answer
 from core.api.filters import UUIDInFilter
 
 
@@ -23,14 +23,4 @@ class AnswerCommentFilterSet(filters.FilterSet):
         model = Answer
         fields = [
             "answer",
-        ]
-
-
-class AnswerCrossCheckFilterSet(filters.FilterSet):
-    question = UUIDInFilter(field_name="answer__question__slug", required=True)
-
-    class Meta:
-        model = AnswerCrossCheck
-        fields = [
-            "question",
         ]
